requirements.txt
agno==1.4.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.1.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.9.1
bleach==5.0.1
blinker==1.6.3
blis==0.7.11
bs4==0.0.1
build==1.0.3
cachetools==5.3.1
catalogue==1.0.2
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chess==1.11.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.1
confection==0.1.3
contourpy==1.0.7
cryptography==41.0.4
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.6.1
docstring_parser==0.16
docx2txt==0.8
elastic-transport==8.4.1
elasticsearch==8.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
entrypoints==0.4
es-indexer==0.1
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.109.2
fastjsonschema==2.16.2
filelock==3.11.0
Flask==3.0.2
Flask-Cors==3.0.9
flatbuffers==23.5.26
fonttools==4.39.3
fsspec==2024.2.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.28.0
google-genai==1.11.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.8
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.18.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.10
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kubernetes==29.0.0
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.1.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==21.3
pafy==0.5.5
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pdfminer.six==20221105
pdfminer3==2018.12.3.0
pickleshare==0.7.5
pillow==10.2.0
plac==1.1.3
platformdirs==2.5.4
plotly==5.17.0
posthog==3.4.1
preshed==3.0.9
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.24.4
psutil==5.9.4
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.13.0
pymongo==4.6.1
PyMySQL==1.1.0
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyresparser==1.0.6
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2022.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.0
six==1.15.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.0.1
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
stack-data==0.6.2
starlette==0.36.3
streamlit==1.27.2
streamlit-tags==1.2.8
sympy==1.11.1
tenacity==8.2.3
terminado==0.17.0
textblob==0.18.0.post0
thinc==7.4.6
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.66.1
traitlets==5.5.0
transformers==4.34.1
typer==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.1
urllib3==1.25.10
uvicorn==0.27.1
validators==0.22.0
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.5
weasel==0.3.2
webencodings==0.5.1
websocket-client==1.4.2
websockets==15.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.3
wrapt==1.16.0
zipp==3.17.0
